backend/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
PyMySQL==1.1.1
sqlparse==0.5.1
zope.event==5.0
zope.interface==7.0.1
